Move topbar logic into hooks and utils
